refactor(front): stale cache helper 분리
